requirements.txt
dagster
dagit
fastapi
uvicorn[standard]
google-generativeai
requests
python-dotenv
dagster-slack